BIG REFACTOR: use tup

+++ code/src/ui_common.h
@@ -5,7 +5,7 @@
 #define HEIGHT 48
 
 extern unsigned int global_clock;
-extern double cadence;
+extern unsigned int cadence;
 
 void setCorners();
 void setBigText(char *s, char x, char y);
